feat(product overview) [VWS4LS] switch viewer button and extract data loading (#209)
* feat(product-viewer): Add ProductOverviewCard and product viewer page

* feat(product-overview): Enhance ProductOverviewCard with additional properties expect for ECLASE, IEC, VEC

* feat(ProductOverviewCard): Refactor data retrieval and enhance submodel element handling

* feat(env-product-view):add product view feature flag

* productclassification

* productclassification

* fix missing parameter

* added switch viewer button

* extracted notFound Component

* extracted AasDataLoader Hook

* extracted UseProductImageUrl Hook Hook

* extracted MobileAccordion Component

* feat(env for product view) (#211)

* feat(env-product-view):add product view feature flag

* feat(env-product-view):navigate to product view from AAS list

* feat(env-product-view):docu for new env

* fix merge

* fix merge

* fix productimage loading

---------
