+++ src/app/[locale]/product/[base64AasId]/page.tsx
@@ -1,142 +1,47 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import { Box, Button, Skeleton, Typography } from '@mui/material';
 import { safeBase64Decode } from 'lib/util/Base64Util';
-import { ArrowForward } from '@mui/icons-material';
-import { Reference, Submodel } from '@aas-core-works/aas-core3.0-typescript/types';
 import { useIsMobile } from 'lib/hooks/UseBreakpoints';
 import { getTranslationText } from 'lib/util/SubmodelResolverUtil';
 import { useParams, useRouter, useSearchParams } from 'next/navigation';
 import { SubmodelsOverviewCard } from 'app/[locale]/viewer/_components/SubmodelsOverviewCard';
 import { ProductOverviewCard } from '../_components/ProductOverviewCard';
 import { useEnv } from 'app/EnvProvider';
-import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
-import {
-    getAasFromRepository,
-    performFullAasSearch,
-    performSubmodelFullSearch,
-} from 'lib/services/search-actions/searchActions';
-import { LocalizedError } from 'lib/util/LocalizedError';
-import {
-    SubmodelOrIdReference,
-    useAasOriginSourceState,
-    useAasState,
-    useRegistryAasState,
-    useSubmodelState,
-} from 'components/contexts/CurrentAasContext';
-import { SubmodelDescriptor } from 'lib/types/registryServiceTypes';
 import { TransferButton } from 'app/[locale]/viewer/_components/transfer/TransferButton';
-import { useShowError } from 'lib/hooks/UseShowError';
-import { useLocale, useTranslations } from 'next-intl';
+import { NoSearchResult } from 'components/basics/detailViewBasics/NoSearchResult';
+import { useAasLoader } from 'lib/hooks/UseAasDataLoader';
+import { useLocale } from 'next-intl';
+import { useTranslations } from 'use-intl';
 
 export default function Page() {
     const navigate = useRouter();
     const searchParams = useParams<{ base64AasId: string }>();
     const base64AasId = decodeURIComponent(searchParams.base64AasId).replace(/=+$|[%3D]+$/, '');
     const aasIdDecoded = safeBase64Decode(base64AasId);
-    const [isLoadingAas, setIsLoadingAas] = useState(false);
     const isMobile = useIsMobile();
     const locale = useLocale();
     const env = useEnv();
     const encodedRepoUrl = useSearchParams().get('repoUrl');
     const repoUrl = encodedRepoUrl ? decodeURI(encodedRepoUrl) : undefined;
-    const [aasOriginUrl, setAasOriginUrl] = useAasOriginSourceState();
-    const [aasFromContext, setAasFromContext] = useAasState();
-    const [submodels, setSubmodels] = useSubmodelState();
-    const [isSubmodelsLoading, setIsSubmodelsLoading] = useState(true);
-    const [registryAasData, setRegistryAasData] = useRegistryAasState();
-    const { showError } = useShowError();
+    // TODO refactor translation strings here
     const t = useTranslations('pages.aasViewer');
+    const tp = useTranslations('pages.productViewer');
 
-
-
-    const submodelWhitelist: string[] = JSON.parse(env.SUBMODEL_WHITELIST || '[]');
-
-    function whitelistContains(sm: Submodel) {
-        const semanticIds = sm.semanticId?.keys.map((key) => key.value);
-        return semanticIds?.some((id) => submodelWhitelist.includes(id));
-    }
-
-    useAsyncEffect(async () => {
-        await fetchSubmodels();
-    }, [aasFromContext]);
-
-    useEffect(() => {
-        return () => {
-            setSubmodels(new Array<SubmodelOrIdReference>());
-        };
-    }, []);
-
-    useAsyncEffect(async () => {
-        if (aasFromContext) {
-            return;
-        }
-        setIsLoadingAas(true);
-        await loadAasContent();
-        setIsLoadingAas(false);
-    }, [base64AasId, env]);
-
-    async function loadAasContent() {
-        if (repoUrl) {
-            const response = await getAasFromRepository(aasIdDecoded, repoUrl);
-            if (response.isSuccess) {
-                setAasOriginUrl(repoUrl);
-                setAasFromContext(response.result);
-                return;
-            }
-        }
-
-        const { isSuccess, result } = await performFullAasSearch(aasIdDecoded);
-        if (!isSuccess) {
-            showError(new LocalizedError('navigation.errors.urlNotFound'));
-        } else if (result.aas) {
-            setAasOriginUrl(result.aasData?.aasRepositoryOrigin ?? null);
-            setRegistryAasData(result.aasData);
-            setAasFromContext(result.aas);
-        } else {
-            navigate.push(result.redirectUrl);
-        }
-    }
-
-    async function fetchSubmodels() {
-        setIsSubmodelsLoading(true);
-        if (aasFromContext?.submodels) {
-            await Promise.all(
-                aasFromContext.submodels.map(async (smRef, i) => {
-                    const newSm = await fetchSingleSubmodel(smRef, registryAasData?.submodelDescriptors?.[i]);
-                    setSubmodels((submodels) => {
-                        const exists = submodels.some((sm) => sm.id === newSm.id);
-                        if (exists) return submodels;
-                        if (env.WHITELIST_FEATURE_FLAG && newSm.submodel && !whitelistContains(newSm.submodel))
-                            return submodels;
-                        return [...submodels, newSm];
-                    });
-                }),
-            );
-        }
-        setIsSubmodelsLoading(false);
-    }
-
-    async function fetchSingleSubmodel(
-        reference: Reference,
-        smDescriptor?: SubmodelDescriptor,
-    ): Promise<SubmodelOrIdReference> {
-        const submodelResponse = await performSubmodelFullSearch(reference, smDescriptor);
-        if (!submodelResponse.isSuccess)
-            return {
-                id: reference.keys[0].value,
-                error: submodelResponse.errorCode,
-            };
-
-        return {
-            id: submodelResponse.result.id,
-            submodel: submodelResponse.result,
-        };
-    }
+    const {
+        aasFromContext,
+        isLoadingAas,
+        aasOriginUrl,
+        submodels,
+        isSubmodelsLoading,
+    } = useAasLoader(base64AasId, repoUrl);
 
     const startComparison = () => {
         navigate.push(`/compare?aasId=${encodeURIComponent(aasIdDecoded)}`);
+    };
+
+    const goToAASView = () => {
+        navigate.push(`/viewer/${searchParams.base64AasId}`);
     };
 
     const pageStyles = {
@@ -194,6 +99,11 @@ export default function Page() {
                             </Button>
                         )}
                         {env.TRANSFER_FEATURE_FLAG && <TransferButton />}
+                        {env.PRODUCT_VIEW_FEATURE_FLAG &&
+                            <Button variant="contained" sx={{ whiteSpace: 'nowrap' }} onClick={goToAASView}>
+                                {tp('actions.toAasView')}
+                            </Button>
+                        }
                     </Box>
                     <ProductOverviewCard
                         aas={aasFromContext}
@@ -208,28 +118,7 @@ export default function Page() {
                     )}
                 </Box>
             ) : (
-                <>
-                    <Typography
-                        variant="h2"
-                        style={{
-                            width: '90%',
-                            margin: '0 auto',
-                            marginTop: '10px',
-                            overflowWrap: 'break-word',
-                            wordBreak: 'break-word',
-                            textAlign: 'center',
-                            display: 'inline-block',
-                        }}
-                    >
-                        {t('messages.noDataFound')}
-                    </Typography>
-                    <Typography color="text.secondary">
-                        {t('messages.noDataFoundFor', { name: safeBase64Decode(base64AasId) })}
-                    </Typography>
-                    <Button variant="contained" startIcon={<ArrowForward />} href="/">
-                        {t('actions.toHome')}
-                    </Button>
-                </>
+                <NoSearchResult base64AasId={safeBase64Decode(base64AasId)} />
             )}
         </Box>
     );
