+++ src/app/[locale]/product/_components/ProductOverviewCard.tsx
@@ -1,7 +1,4 @@
 import {
-    Accordion,
-    AccordionDetails,
-    AccordionSummary,
     Box,
     Card,
     CardContent,
@@ -10,19 +7,14 @@ import {
 } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import { DataRow } from 'components/basics/DataRow';
-import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { AssetAdministrationShell, SubmodelElementCollection } from '@aas-core-works/aas-core3.0-typescript/types';
 import { IconCircleWrapper } from 'components/basics/IconCircleWrapper';
 import { AssetIcon } from 'components/custom-icons/AssetIcon';
 import { ShellIcon } from 'components/custom-icons/ShellIcon';
-import { isValidUrl } from 'lib/util/UrlUtil';
 import { encodeBase64 } from 'lib/util/Base64Util';
-import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
 import { useRouter } from 'next/navigation';
 import { SubmodelOrIdReference, useAasState } from 'components/contexts/CurrentAasContext';
 import { ImageWithFallback } from 'components/basics/StyledImageWithFallBack';
-import { getThumbnailFromShell } from 'lib/services/repository-access/repositorySearchActions';
-import { mapFileDtoToBlob } from 'lib/util/apiResponseWrapper/apiResponseWrapper';
 import { useLocale, useTranslations } from 'next-intl';
 import { SubmodelSemanticIdEnum } from 'lib/enums/SubmodelSemanticId.enum';
 import {
@@ -30,7 +22,9 @@ import {
     findSubmodelElementByIdShort,
     findValueByIdShort,
 } from 'lib/util/SubmodelResolverUtil';
+import { MobileAccordion } from 'components/basics/detailViewBasics/MobileAccordion';
 import { SubmodelElementSemanticIdEnum } from 'lib/enums/SubmodelElementSemanticId.enum';
+import { useProductImageUrl } from 'lib/hooks/UseProductImageUrl';
 
 type ProductOverviewCardProps = {
     readonly aas: AssetAdministrationShell | null;
@@ -40,12 +34,6 @@ type ProductOverviewCardProps = {
     readonly isAccordion: boolean;
     readonly imageLinksToDetail?: boolean;
     readonly repositoryURL: string | null;
-};
-
-type MobileAccordionProps = {
-    readonly content: React.ReactNode;
-    readonly title: string;
-    readonly icon: React.ReactNode;
 };
 
 type ProductClassification = {
@@ -60,59 +48,18 @@ type OverviewData = {
     readonly manufacturerProductFamily?: string;
     readonly manufacturerProductType?: string;
     readonly productClassifications?: ProductClassification[];
-    //Hier noch mehr Properties hinzufügen
 };
-
-function MobileAccordion(props: MobileAccordionProps) {
-    return (
-        <Accordion disableGutters elevation={0} style={{ width: '100%' }}>
-            <AccordionSummary expandIcon={<ArrowDropDownIcon sx={{ color: 'grey.600' }} />}>
-                <Box display="flex" alignItems="center" data-testid="mobile-accordion-header">
-                    <IconCircleWrapper sx={{ mr: 1 }}>{props.icon}</IconCircleWrapper>
-                    <Typography>{props.title}</Typography>
-                </Box>
-            </AccordionSummary>
-            <AccordionDetails data-testid="mobile-accordion-content">{props.content}</AccordionDetails>
-        </Accordion>
-    );
-}
 
 export function ProductOverviewCard(props: ProductOverviewCardProps) {
     const isAccordion = props.isAccordion;
     const navigate = useRouter();
-    const [productImageUrl, setProductImageUrl] = useState<string>('');
     const [, setAasState] = useAasState();
     const t = useTranslations('pages.productViewer');
     const [overviewData, setOverviewData] = useState<OverviewData>();
     const locale = useLocale();
+    const productImageUrl = useProductImageUrl(props.aas, props.repositoryURL, props.productImage);
 
-    async function createAndSetUrlForImageFile() {
-        if (!props.aas) return;
-
-        if (!props.repositoryURL) {
-            setProductImageUrl('');
-            return;
-        }
-
-        const response = await getThumbnailFromShell(props.aas.id, props.repositoryURL);
-        if (!response.isSuccess) {
-            console.error('Image not found');
-            return;
-        }
-        const blob = mapFileDtoToBlob(response.result);
-        setProductImageUrl(URL.createObjectURL(blob));
-    }
-
-    useAsyncEffect(async () => {
-        if (!props.productImage) return;
-
-        if (!isValidUrl(props.productImage!)) {
-            await createAndSetUrlForImageFile();
-        } else {
-            setProductImageUrl(props.productImage);
-        }
-    }, [props.productImage]);
-
+    // TODO: once we know how to display this data, split it into its own components
     useEffect(() => {
         if (props.submodels && props.submodels.length > 0) {
             const technicalDataSubmodelElements = findSubmodelByIdOrSemanticId(
