+++ src/locale/de.json
@@ -96,6 +96,11 @@
     "dummyComponent": {
       "dummyText": "Das ist eine Dummy-Komponente. Sie wird verwendet, um das Layout und Design der Anwendung zu testen."
     },
+    "noSearchResult": {
+      "header": "Es wurden keine Daten für \"{name}\" gefunden.",
+      "noDataFound": "Es wurden keine Daten für \"{name}\" gefunden.",
+      "toHomeButton": "Zur Startseite"
+    },
     "qrScanner": {
       "errors": {
         "defaultCallbackErrorMsg": "QR Code konnte nicht geöffnet werden!",
@@ -195,8 +200,8 @@
     },
     "aasViewer": {
       "actions": {
-        "toHome": "Zur Startseite",
-        "compareButton": "Vergleichen"
+        "compareButton": "Vergleichen",
+        "toProductView": "Produktansicht"
       },
       "asset": "Asset",
       "assetAdministrationShell": "Verwaltungsschale",
@@ -211,10 +216,6 @@
         "notAvailable": "-",
         "text": "Text",
         "value": "Wert"
-      },
-      "messages": {
-        "noDataFoundFor": "Es wurden keine Daten für \"{name}\" gefunden.",
-        "noDataFound": "Keine Daten gefunden"
       },
       "submodels": {
         "actions": {
@@ -285,6 +286,9 @@
       "classification": "Produktklassifikation",
       "error":{
         "technicalDataSubmodelNotFound": "Technical Data Submodel nicht gefunden"
+      },
+      "actions": {
+        "toAasView": "AAS Ansicht"
       }
     },
     "settings": {
